+++ polyscan/testdata/godeps/app/extra.go
@@ -1,0 +1,5 @@
+package main
+
+import "example.com/godeps/lib"
+
+var _ = lib.Count